✨ Add new structure for the configs structs
